GUI: The "General" page of the configuration dialog has been changed. Instead of LCD numbers, there are now spin boxes.
There are also some behind-the-scenes modifications to the layout system to make it easier to add/remove config elements at a later date, but it doesn't change the view at all.

I'm working on making the same changes to the "Game" page, but it's a bit more complicated than the "General" page.

svn path=/trunk/KDE/kdegames/kspaceduel/; revision=637567